# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.Order:
    properties:
      customer_name:
        type: string
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/handler.OrderItem'
        type: array
      status:
        type: string
      table_id:
        type: integer
      total_amount:
        type: number
    type: object
  handler.OrderItem:
    properties:
      id:
        type: integer
      menu_item_id:
        type: integer
      order_id:
        type: integer
      quantity:
        type: integer
      special_instructions:
        type: string
    type: object
  handler.OrderResponse:
    properties:
      customer_name:
        type: string
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/handler.OrderItem'
        type: array
      restaurant_id:
        type: integer
      restaurant_name:
        type: string
      status:
        type: string
      table_id:
        type: integer
      total_amount:
        type: number
    type: object
  handler.OrderStatusUpdate:
    properties:
      status:
        example: completed
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API for Order System with user authentication and restaurant management
  title: Order System API
  version: "1.0"
paths:
  /api/order:
    get:
      description: Get all orders for all restaurants belonging to the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.OrderResponse'
            type: array
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Error retrieving orders
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all user orders
      tags:
      - Order
...
  /api/restaurant/{restaurant_id}/order:
    get:
      description: Get all orders for a restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Order'
            type: array
        "404":
          description: Restaurant not found
          schema:
            type: string
        "500":
          description: Error retrieving orders
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all orders
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: Create a new order for a restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/handler.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Order'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Restaurant, table, or menu item not found
          schema:
            type: string
        "500":
          description: Error creating order
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - Order
  /api/restaurant/{restaurant_id}/order/{id}:
    delete:
      description: Delete an order
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Restaurant or order not found
          schema:
            type: string
        "500":
          description: Error deleting order
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete an order
      tags:
      - Order
    get:
      description: Get a single order by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Order'
        "404":
          description: Restaurant or order not found
          schema:
            type: string
        "500":
          description: Error retrieving order
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Order
    patch:
      consumes:
      - application/json
      description: Update the status of an order
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Order status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/handler.OrderStatusUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Order'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Restaurant or order not found
          schema:
            type: string
        "500":
          description: Error updating order
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - Order
...
  /api/restaurants/{restaurant_id}/order:
    post:
      consumes:
      - application/json
      description: Create a new order for a restaurant without authentication
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/handler.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Order'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Restaurant, table, or menu item not found
          schema:
            type: string
        "500":
          description: Error creating order
          schema:
            type: string
      summary: Create a public order
      tags:
      - Order
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"